1.chebase中加入了MAC_OS_X下面CHE_Lock的实现 2.MAC_OS_X下面页面渲染开始对tiling进行支持 3.工程文件的改变 4.一些函数声明的修改
git-svn-id: https://10.211.55.3/svn/chelib@809 5bab769a-80ee-f445-9178-c608493bda25

+++ trunk/include/che_define.h
@@ -77,7 +77,7 @@ typedef int  HE_INTPTR;
 typedef unsigned int HE_UINTPTR;
 
 //32-bit floating-point number.
-typedef float  HE_FLOAT;
+typedef double  HE_FLOAT;
 
 typedef double HE_DOUBLE;
